feat: добавляет модульность
